requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-Environments==0.1
Flask-Injector==0.4.0
Jinja2==2.7.3
MarkupSafe==0.23
PyHamcrest==1.8.0
PyYAML==3.11
Pygments==1.6
Sphinx==1.2.2
Werkzeug==0.9.6
argh==0.25.0
argparse==1.2.1
behave==1.2.4
click==3.2
docutils==0.12
enum34==1.0
injector==0.9.0
itsdangerous==0.24
nose==1.3.3
nose-watch==0.9.1
parse==1.6.4
parse-type==0.3.4
pathtools==0.1.2
pymongo==2.7.2
simplejson==3.6.0
six==1.7.3
watchdog==0.8.1
webassets==0.10.1
wsgiref==0.1.2
